IGNITE-7612 Web Console: Refactored mongoose schemas to separate module.
(cherry picked from commit 71db707)
